+++ cognite/powerops/resync/models/cogshop.py
@@ -41,6 +41,8 @@ class CogShop1Asset(CogShopCore, DataModel, protected_namespaces=()):
         ContainerId("cogShop", "Transformation"): cogshop_v1.TransformationApply,
     }
     model_templates: dict[ExternalID, cogshop_v1.ModelTemplateApply] = Field(default_factory=dict)
+    mappings: dict[ExternalID, cogshop_v1.MappingApply] = Field(default_factory=dict)
+    transformations: dict[ExternalID, cogshop_v1.TransformationApply] = Field(default_factory=dict)
     base_mappings: list[CDFSequence] = Field(default_factory=list)
     output_definitions: list[CDFSequence] = Field(default_factory=list)
 
@@ -48,21 +50,21 @@ class CogShop1Asset(CogShopCore, DataModel, protected_namespaces=()):
     def from_cdf(cls, client: PowerOpsClient, data_set_external_id: str) -> CogShop1Asset:
         cog_shop = client.cog_shop1
         templates = cog_shop.model_templates.list(limit=-1)
-        base_mapping_ids = list({mapping for t in templates for mapping in t.base_mappings})
-        base_mappings = cog_shop.mappings.retrieve(base_mapping_ids)
-        transformations_ids = list({t for m in base_mappings for t in m.transformations})
-        transformations = cog_shop.transformations.retrieve(transformations_ids)
-        file_ids = list({t.model for t in templates})
-        files = cog_shop.file_refs.retrieve(file_ids)
+        base_mappings = cog_shop.mappings.list(limit=-1)
+        transformations = cog_shop.transformations.list(limit=-1)
+        files = cog_shop.file_refs.list(limit=-1)
 
         transformation_by_id = {t.external_id: t for t in transformations}
         mappings_by_id = {}
         readme_fields = {"created_time", "deleted_time", "last_updated_time", "version"}
+        for transformation in transformations:
+            data = transformation.model_dump(exclude=readme_fields)
+            apply = cogshop_v1.TransformationApply(**data)
+            transformation_by_id[apply.external_id] = apply
+
         for mapping in base_mappings:
             data = mapping.model_dump(exclude=readme_fields)
-            data["transformations"] = [
-                transformation_by_id[t].model_dump(exclude=readme_fields) for t in data["transformations"]
-            ]
+            data["transformations"] = [transformation_by_id[t] for t in data["transformations"]]
             apply = cogshop_v1.MappingApply(**data)
             mappings_by_id[apply.external_id] = apply
 
@@ -101,6 +103,8 @@ class CogShop1Asset(CogShopCore, DataModel, protected_namespaces=()):
 
         return cls(
             model_templates=model_templates,
+            mappings=mappings_by_id,
+            transformations=transformation_by_id,
             base_mappings=base_mappings,
             output_definitions=output_definitions,
             shop_files=shop_files,